Removed send_sources
